backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
redis
psycopg2-binary
alembiccelery